==> 02_fitting-exp2-five-objects.qmd <==
---
title: "Fitting ideal observer models. Experiment 2: five game objects."
format: html
---

```{r}
#| warning: false
#| message: false
library(tidyverse)

library(cmdstanr)
library(glue)

FORCE_RESAMPLE <- FALSE
```

## Import proprocessed data
```{r}
data_for_stan <- readRDS("Data/exp2-five-objects/exp2-standata.RDS")
```


## Compiling the models
```{r}
model_names <- c("[set-none]_[obj-common]_[color-none]",
                 "[set-none]_[obj-perset]_[color-none]",
                 
                 
                 
                 "[set-same]_[obj-common]_[color-none]",
                 "[set-same]_[obj-common]_[color-objboost]",
                 "[set-same]_[obj-common]_[color-common]",
                 "[set-same]_[obj-common]_[color-perset]",
                 
                 "[set-same]_[obj-perset]_[color-none]",
                 "[set-same]_[obj-perset]_[color-objboost]",
                 "[set-same]_[obj-perset]_[color-common]",
                 "[set-same]_[obj-perset]_[color-perset]",
                 
                 
                 
                 "[set-different]_[obj-common]_[color-none]",
                 "[set-different]_[obj-common]_[color-objboost]",
                 "[set-different]_[obj-common]_[color-common]",
                 "[set-different]_[obj-common]_[color-perset]",
                 
                 "[set-different]_[obj-perset]_[color-none]",
                 "[set-different]_[obj-perset]_[color-objboost]",
                 "[set-different]_[obj-perset]_[color-common]",
                 "[set-different]_[obj-perset]_[color-perset]",
                 

                 "[set-memory]_[obj-common]_[color-none]",      
                 "[set-memory]_[obj-common]_[color-objboost]",
                 "[set-memory]_[obj-common]_[color-common]",
                 "[set-memory]_[obj-common]_[color-perset]",

                 "[set-memory]_[obj-perset]_[color-none]",
                 "[set-memory]_[obj-perset]_[color-objboost]",
                 "[set-memory]_[obj-perset]_[color-common]",
                 "[set-memory]_[obj-perset]_[color-perset]"
                 )

models <- purrr::map(model_names, ~cmdstan_model(glue("Stan/exp23/{.}.stan"),
                                                 exe_file = glue("Stan/exp23/exe/{.}.stan")))
names(models) <- model_names
```

## Fitting models
```{r}
for(im in 1:length(model_names)) {
  cat(model_names[im], "\n")
  loo_filename <- fs::path("Loos/exp2-five-objects", glue("loo_exp2_{model_names[im]}.RDS"))
  loglik_filename <- fs::path("LogLik/exp2-five-objects", glue("loglik_exp2_{model_names[im]}.RDS"))
  draws_filename <- fs::path("Draws/exp2-five-objects", glue("draws_exp2_{model_names[im]}.RDS")) 
  
  if (fs::file_exists(loo_filename) & fs::file_exists(draws_filename) & !FORCE_RESAMPLE){
    cat("    Already sampled, skipping...\n")
    next
  } 
  
  fit <- models[[im]]$sample(data = data_for_stan,
                             chains = 4,
                             parallel_chains = future::availableCores())

  # loo
  fit_loo <- fit$loo()
  saveRDS(fit_loo, loo_filename)
  
  # log_lik
  log_lik <- fit$draws(format = "df") |>
              as_tibble() |>
              select(starts_with('.'), starts_with("log_lik"))
  saveRDS(log_lik, loglik_filename)

  # draws
  draws <- fit$draws(format = "df") |>
           as_tibble() |>
           select(-starts_with("log_lik"))

  saveRDS(draws, draws_filename)
}
```
